보호자 페이지, ToggleForm 컴포넌트 재분리
